requirements.txt
Flask==3.0.2
Werkzeug==3.0.1
python-pptx==0.6.23
gTTS==2.5.1
moviepy
pydub==0.25.1
numpy==1.26.4
Pillow==10.2.0
opencv-python==4.9.0.80
ffmpeg-python==0.2.0
uuid==1.30
python-dotenv
elevenlabs